app/core/scraping.py
"""Price scraping module for SmartNinja application.

This module provides implementations for scraping phone prices
from various sources using the Bright Data API service.
"""

import os
from datetime import datetime
from typing import Dict, List

from dotenv import load_dotenv

load_dotenv()
BRIGHT_DATA_KEY = os.getenv("BRIGHT_DATA_KEY")


# pylint: disable=too-few-public-methods
class PriceScraper:
    """Service for scraping mobile phone prices"""

    def __init__(self):
        self.base_url = "https://api.brightdata.com"
        self.headers = {"Authorization": f"Bearer {BRIGHT_DATA_KEY}"}

    def get_price_data(self, phone_model: str, region: str) -> Dict:
        """
        Get price data for a specific phone model and region
        Args:
            phone_model: The phone model to scrape
            region: The region to scrape from
        Returns:
            Dictionary containing price data
        """
        try:
            # This is a placeholder - in production, this would use Bright Data's API
            # to scrape real prices from various e-commerce sites
            return {
                "model": phone_model,
                "region": region,
                "price": self._get_mock_price(phone_model, region),
                "currency": "USD",
                "timestamp": datetime.utcnow().isoformat(),
                "source": "mock_data",
            }
        except ValueError as validation_error:
            # Uso de exceção específica em vez de Exception genérica
            raise ValueError(
                f"Invalid input parameters: {str(validation_error)}"
            ) from validation_error
        except RuntimeError as runtime_error:
            # Uso de exceção específica em vez de Exception genérica
            raise RuntimeError(
                f"Runtime error during scraping: {str(runtime_error)}"
            ) from runtime_error

    def _get_mock_price(self, phone_model: str, region: str) -> float:
        """Generate mock price data for demonstration"""
        base_prices = {
            "iPhone 15 Pro": 1000,
            "Samsung Galaxy S24 Ultra": 1200,
            "Google Pixel 8 Pro": 900,
        }
        # Add some regional variation
        price_variation = {"US": 1.0, "EU": 1.1, "BR": 1.2, "IN": 0.9}
        return base_prices.get(phone_model, 1000) * price_variation.get(region, 1.0)


class PriceAnalyzer:
    """Service for analyzing price data with AI using Ollama"""

    def __init__(self):
        self.api_url = os.getenv("OLLAMA_API_URL", "http://localhost:11434")
        self.model = os.getenv("OLLAMA_MODEL", "llama3")
        self.timeout = int(os.getenv("OLLAMA_TIMEOUT", "30"))

    def analyze_market(self, price_data: List[Dict]) -> str:
        """
        Analyze market data using Ollama AI
        Args:
            price_data: List of price data dictionaries
        Returns:
            AI-generated analysis
        """
        try:
            # Format price data for the prompt
            price_info = "\n".join(
                [
                    f"- {item['model']} in {item['region']}: ${item.get('price', 'N/A')}"
                    for item in price_data
                ]
            )
            # Prepare prompt for Ollama
            prompt = f"""Analyze the following mobile phone price data and provide insights:
            
            {price_info}
            
            Please provide a structured analysis including price trends, opportunities and recommendations."""
            # In production, this would call the Ollama API
            # headers = {"Content-Type": "application/json"}
            # payload = {"model": self.model, "prompt": prompt}
            # response = requests.post(f"{self.api_url}/api/generate", json=payload, headers=headers, timeout=self.timeout)
            # if response.status_code == 200:
            #     return response.json().get("response", self._get_fallback_analysis())
            # return self._get_fallback_analysis()
            # For development/testing, return mock analysis
            return self._get_fallback_analysis()
        except Exception as error:
            print(f"Error analyzing price data with Ollama: {str(error)}")
            return self._get_fallback_analysis()

    def _get_fallback_analysis(self) -> str:
        """Provide a fallback analysis when Ollama is unavailable"""
        return f"""
        Market Analysis for {datetime.now().strftime('%Y-%m-%d')}:
        
        1. Price Trends:
        - Overall market is showing stability
        - Regional variations are within expected ranges
        
        2. Opportunities:
        - No significant price drops detected
        - Consider monitoring for future opportunities
        
        3. Recommendations:
        - Continue monitoring for regional price discrepancies
        - Watch for upcoming product launches
        """
