requirements.txt
# Web UI framework dependencies
streamlit==1.29.0

# Data handling
pandas==2.1.0
numpy==1.26.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter
# SQLite is included in Python's standard library

# External services
requests==2.31.0
python-dotenv==1.0.0
brightdata==0.0.2  # Web scraping service
openai==1.3.0      # AI/LLM service

# Data visualization
plotly==5.19.0

# Machine Learning
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
faker==20.1.0       # Test data generation

# Development tools
black==23.11.0      # Code formatting
isort==5.12.0       # Import sorting
pylint==2.17.1      # Linting
bandit==1.7.5       # Security checks

# Production
gunicorn==21.2.0    # Production WSGI server
